requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==3.2.0
cffi==1.17.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
ecdsa==0.19.0
fastapi[standard]

greenlet==3.0.3
h11==0.14.0
idna==3.7
mysqlclient==2.2.4
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
typing_extensions==4.12.2
uvicorn==0.30.6
